Move client state unit tests back to ibc-clients
